Removed duplicate code of palex & lexergen - also seperated the command line interface into an extra file
